hw-bundler/pyproject.toml
[tool.poetry]
name = "hw-bundler"
version = "0.1.0"
description = "Bundles homework to make turning it in easier"
authors = ["Sean Lickfold <seansbox@gmail.com>"]

[tool.poetry.dependencies]
python = "<3.11,>=3.6"
python-docx = "^0.8.11"

[tool.poetry.dev-dependencies]
pyinstaller = "^4.10"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
